requirements.txt
alembic==1.16.2
APScheduler==3.11.0
black==25.1.0
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
Faker==37.3.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
fpdf==1.7.2
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
qrcode==8.2
six==1.17.0
SQLAlchemy==2.0.41
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
WTForms==3.2.1
